requirements.txt
streamlit
numpy
imageio
scipy
opencv-python-headless
Pillow